+++ Sources/Tabman/Bar/BarButton/Types/TMTabItemBarButton.swift
@@ -19,16 +19,19 @@ open class TMTabItemBarButton: TMBarButton {
         static let labelPadding: CGFloat = 4.0
         static let labelTopPadding: CGFloat = 6.0
         static let shrunkenImageScale: CGFloat = 0.9
-        static let badgeInsets = UIEdgeInsets(top: 4.0, left: 0.0, bottom: 0.0, right: 4.0)
+        static let badgeInsets = UIEdgeInsets(top: 4.0, left: 4.0, bottom: 0.0, right: 4.0)
     }
     
     // MARK: Properties
     
+    private let container = UIView()
     private let label = AnimateableLabel()
     private let imageView = UIImageView()
     
     private var imageWidth: NSLayoutConstraint!
     private var imageHeight: NSLayoutConstraint!
+    
+    private var componentConstraints: [NSLayoutConstraint]?
     
     // MARK: Customization
     
@@ -60,7 +63,7 @@ open class TMTabItemBarButton: TMBarButton {
         }
     }
     /// Font of the text label.
-    open var font: UIFont = UIFont.systemFont(ofSize: 12.0, weight: .medium) {
+    open var font: UIFont! {
         didSet {
             label.font = font
         }
@@ -92,25 +95,14 @@ open class TMTabItemBarButton: TMBarButton {
     open override func layout(in view: UIView) {
         super.layout(in: view)
         
-        label.textAlignment = .center
+        view.addSubview(container)
+        container.addSubview(imageView)
+        container.addSubview(label)
         
-        view.addSubview(imageView)
-        view.addSubview(label)
-        
+        container.translatesAutoresizingMaskIntoConstraints = false
         imageView.translatesAutoresizingMaskIntoConstraints = false
         label.translatesAutoresizingMaskIntoConstraints = false
         
-        // core layout
-        let constraints = [
-            imageView.topAnchor.constraint(equalTo: view.topAnchor, constant: Defaults.imagePadding),
-            imageView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            imageView.leadingAnchor.constraint(greaterThanOrEqualTo: view.leadingAnchor, constant: Defaults.imagePadding),
-            label.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: Defaults.labelTopPadding),
-            label.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: Defaults.labelPadding),
-            view.trailingAnchor.constraint(equalTo: label.trailingAnchor, constant: Defaults.labelPadding),
-            label.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-        ]
-        NSLayoutConstraint.activate(constraints)
         label.setContentHuggingPriority(.defaultLow, for: .horizontal)
         
         // set image size
@@ -121,8 +113,9 @@ open class TMTabItemBarButton: TMBarButton {
         
         selectedTintColor = tintColor
         tintColor = .black
-        label.font = self.font
+        font = defaultFont(for: .current)
         label.text = "Item"
+        label.textAlignment = .center
     }
     
     open override func layoutBadge(_ badge: TMBadgeView, in view: UIView) {
@@ -132,8 +125,7 @@ open class TMTabItemBarButton: TMBarButton {
         view.addSubview(badge)
         badge.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
-            badge.topAnchor.constraint(equalTo: view.topAnchor, constant: insets.top),
-            view.trailingAnchor.constraint(equalTo: badge.trailingAnchor, constant: insets.right)
+            badge.topAnchor.constraint(equalTo: view.topAnchor, constant: insets.top)
             ])
     }
     
@@ -155,6 +147,128 @@ open class TMTabItemBarButton: TMBarButton {
         if shrinksImageWhenUnselected {
             let interpolatedScale = 1.0 - ((1.0 - selectionState.rawValue) * (1.0 - Defaults.shrunkenImageScale))
             imageView.transform = CGAffineTransform(scaleX: interpolatedScale, y: interpolatedScale)
+        }
+    }
+    
+    // MARK: Layout
+    
+    open override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+        
+        makeComponentConstraints(for: UIDevice.current.orientation)
+    }
+    
+    private func makeComponentConstraints(for orientation: UIDeviceOrientation) {
+        guard let parent = container.superview else {
+            return
+        }
+        NSLayoutConstraint.deactivate(componentConstraints ?? [])
+        
+        let constraints: [NSLayoutConstraint]
+        
+        // If landscape or we are a `.regular` size class
+        // Translates to:  Landscape || iPad
+        // Tab views will be aligned horizontally.
+        if orientation.isLandscape || traitCollection.horizontalSizeClass == .regular {
+            
+            let imagePadding = traitCollection.horizontalSizeClass == .compact ? Defaults.imagePadding / 2 : Defaults.imagePadding
+            let labelPadding = Defaults.labelPadding
+            
+            constraints = makeHorizontalAlignedConstraints(in: parent,
+                                                           imagePadding: imagePadding,
+                                                           labelPadding: labelPadding)
+        } else { // Default (Portrait on compact) - Vertical alignment
+            
+            let imagePadding = Defaults.imagePadding
+            let labelPadding =  Defaults.labelPadding
+            
+            constraints = makeVerticalAlignedConstraints(in: parent,
+                                                         imagePadding: imagePadding,
+                                                         labelPadding: labelPadding)
+        }
+        
+        componentConstraints = constraints
+        NSLayoutConstraint.activate(constraints)
+    }
+    
+    private func makeHorizontalAlignedConstraints(in parent: UIView,
+                                                  imagePadding: CGFloat,
+                                                  labelPadding: CGFloat) -> [NSLayoutConstraint] {
+        var constraints = [NSLayoutConstraint]()
+        
+        // Container
+        constraints.append(contentsOf: [
+            container.leadingAnchor.constraint(greaterThanOrEqualTo: parent.leadingAnchor),
+            container.topAnchor.constraint(greaterThanOrEqualTo: parent.topAnchor),
+            parent.trailingAnchor.constraint(greaterThanOrEqualTo: container.trailingAnchor),
+            parent.bottomAnchor.constraint(greaterThanOrEqualTo: container.bottomAnchor),
+            container.centerXAnchor.constraint(equalTo: parent.centerXAnchor),
+            container.centerYAnchor.constraint(greaterThanOrEqualTo: parent.centerYAnchor)
+            ])
+        
+        // Label / Image
+        constraints.append(contentsOf: [
+            imageView.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: imagePadding),
+            imageView.topAnchor.constraint(equalTo: container.topAnchor, constant: imagePadding),
+            container.bottomAnchor.constraint(equalTo: imageView.bottomAnchor, constant: imagePadding),
+            label.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: labelPadding),
+            label.topAnchor.constraint(greaterThanOrEqualTo: container.topAnchor, constant: labelPadding),
+            container.trailingAnchor.constraint(equalTo: label.trailingAnchor, constant: labelPadding),
+            container.bottomAnchor.constraint(greaterThanOrEqualTo: label.bottomAnchor, constant: labelPadding),
+            label.centerYAnchor.constraint(equalTo: container.centerYAnchor)
+            ])
+        
+        // Badge
+        let badgeInsets = Defaults.badgeInsets
+        constraints.append(contentsOf: [
+            badge.leadingAnchor.constraint(equalTo: container.trailingAnchor, constant: badgeInsets.left)
+            ])
+        
+        return constraints
+    }
+    
+    private func makeVerticalAlignedConstraints(in parent: UIView,
+                                                imagePadding: CGFloat,
+                                                labelPadding: CGFloat) -> [NSLayoutConstraint] {
+        var constraints = [NSLayoutConstraint]()
+        
+        // Container
+        constraints.append(contentsOf: [
+            container.leadingAnchor.constraint(equalTo: parent.leadingAnchor),
+            container.topAnchor.constraint(greaterThanOrEqualTo: parent.topAnchor),
+            parent.trailingAnchor.constraint(equalTo: container.trailingAnchor),
+            parent.bottomAnchor.constraint(equalTo: container.bottomAnchor)
+            ])
+        
+        // Label / Image
+        constraints.append(contentsOf: [
+            imageView.topAnchor.constraint(equalTo: container.topAnchor, constant: imagePadding),
+            imageView.centerXAnchor.constraint(equalTo: container.centerXAnchor),
+            imageView.leadingAnchor.constraint(greaterThanOrEqualTo: container.leadingAnchor, constant: imagePadding),
+            label.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: Defaults.labelTopPadding),
+            label.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: labelPadding),
+            container.trailingAnchor.constraint(equalTo: label.trailingAnchor, constant: labelPadding),
+            label.bottomAnchor.constraint(equalTo: container.bottomAnchor)
+            ])
+        
+        // Badge
+        let badgeInsets = Defaults.badgeInsets
+        constraints.append(contentsOf: [
+            parent.trailingAnchor.constraint(equalTo: badge.trailingAnchor, constant: badgeInsets.right)
+            ])
+        
+        return constraints
+    }
+}
+
+extension TMTabItemBarButton {
+    
+    private func defaultFont(for device: UIDevice) -> UIFont {
+        switch device.userInterfaceIdiom {
+        case .pad:
+            return UIFont.systemFont(ofSize: 14.0, weight: .medium)
+        default:
+            return UIFont.systemFont(ofSize: 10.0, weight: .medium)
         }
     }
 }