+++ Sources/Tabman/Bar/BarItem/TMBarItem.swift
@@ -26,6 +26,12 @@ public protocol TMBarItemable: class {
     
     /// Badge value to display.
     var badgeValue: String? { get set }
+
+    /// Returns a short description of the button.
+    var accessibilityLabel: String? { get set }
+
+    /// A brief description of the result of performing an action on the accessibility element, in a localized string.
+    var accessibilityHint: String? { get set }
     
     /// Inform the bar that the item has been updated.
     ///
@@ -43,27 +49,61 @@ extension TMBarItemable {
     }
 }
 
+/// :nodoc:
+extension TMBarItemable {
+    
+    //swiftlint:disable unused_setter_value
+
+    public var accessibilityLabel: String? {
+        set {}
+        get {
+            return nil
+        }
+    }
+    
+    public var accessibilityHint: String? {
+        set {}
+        get {
+            return nil
+        }
+    }
+}
+
 /// Default `TMBarItemable` that can be displayed in a `TMBar`.
-public final class TMBarItem: TMBarItemable {
+open class TMBarItem: TMBarItemable {
     
     // MARK: Properties
     
-    public var title: String? {
+    open var title: String? {
         didSet  {
             setNeedsUpdate()
         }
     }
-    public var image: UIImage?  {
+    open var image: UIImage?  {
         didSet {
             setNeedsUpdate()
         }
     }
     
-    public var badgeValue: String? {
+    open var badgeValue: String? {
         didSet {
             setNeedsUpdate()
         }
     }
+
+    public var accessibilityLabel: String? {
+        didSet {
+            setNeedsUpdate()
+        }
+    }
+
+    public var accessibilityHint: String? {
+        didSet {
+            setNeedsUpdate()
+        }
+    }
+
+    public var isAccessibilityElement: Bool { return true }
         
     // MARK: Init