+++ Sources/Tabman/TabmanViewController.swift
@@ -21,12 +21,14 @@ open class TabmanViewController: PageboyViewController, PageboyViewControllerDel
     ///
     /// - top: Pin to the top of the safe area.
     /// - bottom: Pin to the bottom of the safe area.
+    /// - navigationItem: Set as a `UINavigationItem.titleView`.
     /// - custom: Add the view to a custom view and provide custom layout.
     ///           If no layout is provided, all edge anchors will be constrained
     ///           to the superview.
     public enum BarLocation {
         case top
         case bottom
+        case navigationItem(item: UINavigationItem)
         case custom(view: UIView, layout: ((UIView) -> Void)?)
     }
     
@@ -103,6 +105,17 @@ open class TabmanViewController: PageboyViewController, PageboyViewControllerDel
         super.viewDidLayoutSubviews()
         
         setNeedsInsetsUpdate()
+    }
+    
+    /// A method for calculating insets that are required to layout content between the bars
+    /// that have been added.
+    ///
+    /// One can override this method with their own implementation for custom bar inset calculation, to
+    /// take advantage of automatic insets updates for all Tabman's pages during its lifecycle.
+    ///
+    /// - Returns: information about required insets for current state.
+    open func calculateRequiredInsets() -> Insets {
+        return Insets.for(tabmanViewController: self)
     }
     
     // MARK: Pageboy
@@ -198,14 +211,13 @@ extension TabmanViewController {
             fatalError("Bar is expected to inherit from UIView")
         }
         #endif
-        guard barView.superview == nil else {
-            fatalError("Bar has already been added to view hierarchy.")
-        }
         
         bar.dataSource = dataSource
         bar.delegate = self
         
-        bars.append(bar)
+        if bars.contains(where: { $0 === bar }) == false {
+            bars.append(bar)
+        }
         layoutView(barView, at: location)
         
         updateBar(bar, to: relativeCurrentPosition, animated: false)
@@ -266,6 +278,8 @@ extension TabmanViewController {
     
     private func layoutView(_ view: UIView,
                             at location: BarLocation) {
+        view.removeFromSuperview()
+        
         switch location {
         case .top:
             topBarContainer.addArrangedSubview(view)
@@ -284,6 +298,13 @@ extension TabmanViewController {
                     view.bottomAnchor.constraint(equalTo: superview.bottomAnchor)
                     ])
             }
+        case .navigationItem(let item):
+            let container = ViewTitleViewContainer(for: view)
+            
+            container.frame = CGRect(x: 0.0, y: 0.0, width: 300, height: 50)
+            container.layoutIfNeeded()
+            
+            item.titleView = container
         }
     }
     
@@ -352,7 +373,7 @@ internal extension TabmanViewController {
     }
     
     func setNeedsInsetsUpdate(to viewController: UIViewController?) {
-        let insets = Insets.for(tabmanViewController: self)
+        let insets = calculateRequiredInsets()
         self.requiredInsets = insets
         
         barLayoutGuideTop?.constant = insets.spec.allRequiredInsets.top