+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 `Tabman` adheres to [Semantic Versioning](http://semver.org/).
 
 #### 2.x Releases
+- `2.4.x` Releases - [2.4.0](#240)
 - `2.3.x` Releases - [2.3.0](#230)
 - `2.2.x` Releases - [2.2.0](#220) | [2.2.1](#221)
 - `2.1.x` Releases - [2.1.0](#210) | [2.1.1](#211) | [2.1.2](#212) | [2.1.3](#213) | [2.1.4](#214)
@@ -27,6 +28,37 @@ All notable changes to this project will be documented in this file.
 - `0.6.x` Releases - [0.6.0](#060) | [0.6.1](#061) | [0.6.2](#062)
 - `0.5.x` Releases - [0.5.0](#050) | [0.5.1](#051) | [0.5.2](#052) | [0.5.3](#053)
 - `0.4.x` Releases - [0.4.0](#040) | [0.4.1](#041) | [0.4.2](#042) | [0.4.3](#043) | [0.4.4](#044) | [0.4.5](#045) | [0.4.6](#046) | [0.4.7](#047) | [0.4.8](#048)
+
+---
+## [2.4.0](https://github.com/uias/Tabman/releases/tag/2.4.0)
+Released on 2019-04-19
+
+#### Added
+- Ability to override automatic inset values via `calculateRequiredInsets()` on `TabmanViewController`.
+     - by [WingedDoom](https://github.com/WingedDoom)
+- [#404](https://github.com/uias/Tabman/issues/404) `adjustsAlphaOnSelection` to `TMBarButton`.
+- [#338](https://github.com/uias/Tabman/issues/338) Ability to hide / show the bar with `TMHidingBar`.
+- [#403](https://github.com/uias/Tabman/issues/403) Accessibility support to `TMBarView`.
+     - by [iv-mexx](https://github.com/iv-mexx)
+- [#403](https://github.com/uias/Tabman/issues/403) Accessibility support to `TMBarItem`.
+     - by [iv-mexx](https://github.com/iv-mexx)
+- [#403](https://github.com/uias/Tabman/issues/403) Accessibility support to `TMBarButton`.
+     - by [iv-mexx](https://github.com/iv-mexx)
+- [#321](https://github.com/uias/Tabman/issues/321) `navigationItem(item:)` to `TabmanViewController.BarLocation` with support for embedding a `TMBar` in a `UINavigationBar`.
+
+#### Updated
+- Make `TMBarItem` `open`.
+- Make `.title` open on `TMBarItem`.
+- Make `.image` open on `TMBarItem`.
+- Make `.badgeValue` open on `TMBarItem`.
+- Make `.items` `open` on `TMBarView`.
+- Make `.dataSource` `open` on `TMBarView`.
+- Make `.delegate` `open` on `TMBarView`.
+- Make `.scrollMode` `open` on `TMBarView`.
+- Make `.fadesContentEdges` `open` on `TMBarView`.
+- Make `.spacing` `open` on `TMBarView`.
+- [#396](https://github.com/uias/Tabman/issues/396) Improved layout adaptability of `TMTabItemBarButton` on iPad and in landscape.
+- [#404](https://github.com/uias/Tabman/issues/404) `TMBarButton` no longer performs default fade transition in `update(for selectionState:)`
 
 ---
 ## [2.3.0](https://github.com/uias/Tabman/releases/tag/2.3.0)